MSK21QAS-7[library] fixed-result-component-name
