+++ src/app/app.component.html
@@ -1,0 +1,1 @@
+<sui-result></sui-result>
