--- src/app/library/components/result/result.component.html
@@ -1,19 +1,0 @@
-<div class="result">
-    <div class="result__wrapper">
-        <div class="result__image">
-            <img class="result__image-ico {{generalCurType}}" src="{{ generalType }}" alt={{generalCurAlt}} />
-        </div>
-        <div class="result__information">
-            <h1 class="result__information-title">{{ title }}</h1>
-            <h2 *ngIf="subtitle" class="result__information-subtitle">{{ subtitle }}</h2>
-        </div>
-        <div class="result__navigation">
-            <button class="result__navigation-btn _general">
-                <a class="result__navigation-link _general" href="{{ hrefGeneral[0] }}">{{ hrefGeneral[1] }}</a>
-            </button>
-            <button *ngIf="hrefAdditional" class="result__navigation-btn">
-                <a class="result__navigation-link" href="{{ hrefAdditional[0] }}">{{ hrefAdditional[1] }}</a>
-            </button>
-        </div>
-    </div>
-</div>
