--- src/app/library/components/result/result.component.less
@@ -1,111 +1,0 @@
-.result{
-    &__wrapper{
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        gap:20px;
-    }
-    &__image{
-        &-ico._success{
-            width:64px;
-            height:64px;
-        }
-        &-ico._info{
-            width:64px;
-            height:64px;
-        }
-        &-ico._warning{
-            width:64px;
-            height:64px;
-        }
-        &-ico._error{
-            width:64px;
-            height:64px;
-        }
-        &-ico._403{
-            width:251px;
-            height:294px;
-        }
-        &-ico._404{
-            width:251px;
-            height:294px;
-        }
-        &-ico._500{
-            width:251px;
-            height:294px;
-        }
-    }
-    &__information{
-        display:flex;
-        flex-direction: column;
-        text-align: center;
-        gap:2px;
-        font-family: Arial, Helvetica, sans-serif;
-        &-title{
-            color: #000000d9;
-            font-size: 24px;
-            line-height: 1.8;
-            font-weight: 300;
-        }
-        &-subtitle{
-            color: #00000073;
-            font-size: 14px;
-            line-height: 1.6;
-            font-weight: 300;
-        }
-    }
-    &__navigation{
-        display:flex;
-        gap:10px;
-        &-btn{
-            line-height: 1.5715;
-            position: relative;
-            display: inline-block;
-            font-weight: 400;
-            white-space: nowrap;
-            text-align: center;
-            background-image: none;
-            border: 1px solid transparent;
-            box-shadow: 0 2px #00000004;
-            cursor: pointer;
-            transition: all .3s cubic-bezier(.645,.045,.355,1);
-            -webkit-user-select: none;
-            -moz-user-select: none;
-            -ms-user-select: none;
-            user-select: none;
-            touch-action: manipulation;
-            height: 32px;
-            padding: 4px 15px;
-            font-size: 14px;
-            border-radius: 2px;
-            background: #fff;
-            border-color: #d9d9d9;
-            &:hover{
-                border-color: #40a9ff;
-            }
-            &._general{
-                background:#1890ff;
-                &:hover{
-                    background:#40a9ff;
-                }
-            }
-        }
-        &-link{
-            text-decoration: none;
-            color: #000000d9;
-            transition: all .3s cubic-bezier(.645,.045,.355,1);
-            &._general{
-                color:white;
-                &:hover{
-                    color:white;
-                }
-            }
-            &:active{
-                text-decoration: none;
-            }
-            &:hover{
-                color:#40a9ff;
-            }
-        }
-    }
-}
