+++ src/app/library/components/sui-slider/slider.component.html
@@ -1,0 +1,7 @@
+<div class="slider">
+    <div class="slider__standart">
+        <input [(ngModel)]="value" class="slider__obj" type="range" name="slider" id="slider" value="0" min="0" max="100" step="1">
+        <label for="slider" id="value">{{value}}</label>
+        <div class="slider__progress-bar"></div>
+    </div>
+</div>
