+++ src/app/library/components/sui-slider/slider.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-slider',
+  templateUrl: './slider.component.html',
+  styleUrls: ['./slider.component.less']
+})
+export class SliderComponent implements OnInit {
+
+  value:number = 0;
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
